# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type CourseType = typeof CourseType[keyof typeof CourseType];


export const CourseType = {
  literature_atar: 'literature_atar',
  english_atar: 'english_atar',
  general_english: 'general_english',
} as const;

export type TaskType = typeof TaskType[keyof typeof TaskType];


export const TaskType = {
  essay: 'essay',
  close_reading: 'close_reading',
  short_answer: 'short_answer',
  creative_writing: 'creative_writing',
} as const;

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface StudentProfile {
  userId: string;
  /**
     * @minimum 7
     * @maximum 12
     */
  yearLevel: number;
  courseType: CourseType;
  /** @nullable */
  school: string | null;
  updatedAt: string;
}

export interface StudentProfileEnvelope {
  profile: StudentProfile | null;
}

export interface UpsertStudentProfileBody {
  /**
     * @minimum 7
     * @maximum 12
     */
  yearLevel: number;
  courseType: CourseType;
  /** @nullable */
  school?: string | null;
}

export interface Text {
  id: number;
  title: string;
  author: string;
  yearLevels: number[];
  courseTypes: CourseType[];
}

export interface TextRef {
  id: number;
  title: string;
}

export interface ContentType {
  id: number;
  slug: string;
  name: string;
  description: string;
}

export interface DocumentSummary {
  id: number;
  title: string;
  slug: string;
  contentType: ContentType;
  yearLevels: number[];
  courseTypes: CourseType[];
  taskTypes: TaskType[];
  status: DocumentStatus;
  texts: TextRef[];
  /** @nullable */
  publishedAt: string | null;
  updatedAt: string;
  isNew: boolean;
}

export type DocumentDetail = DocumentSummary & {
  body: string;
  copyrightNotice: string;
};

export interface CreateDocumentBody {
  /** @minLength 1 */
  title: string;
  contentTypeId: number;
  body: string;
  yearLevels: number[];
  courseTypes: CourseType[];
  taskTypes: TaskType[];
  textIds: number[];
  status?: DocumentStatus;
}

export interface UpdateDocumentBody {
  /** @minLength 1 */
  title?: string;
  contentTypeId?: number;
  body?: string;
  yearLevels?: number[];
  courseTypes?: CourseType[];
  taskTypes?: TaskType[];
  textIds?: number[];
  status?: DocumentStatus;
}

export interface DocumentVersion {
  id: number;
  documentId: number;
  versionNumber: number;
  body: string;
  createdAt: string;
}

export interface MetalanguageTerm {
  id: number;
  term: string;
  definition: string;
  tab: string;
  /** @nullable */
  example: string | null;
}

export interface AdminStatus {
  isAdmin: boolean;
}

export interface ProgressRecord {
  documentId: number;
  /** @nullable */
  viewedAt: string | null;
  /** @nullable */
  completedAt: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListTextsParams = {
yearLevel?: number;
courseType?: CourseType;
};

export type ListDocumentsParams = {
yearLevel?: number;
courseType?: CourseType;
textId?: number;
contentTypeId?: number;
taskType?: TaskType;
search?: string;
};

export type ListMetalanguageTermsParams = {
tab?: string;
search?: string;
};

export type AdminListDocumentsParams = {
status?: DocumentStatus;
contentTypeId?: number;
search?: string;
};