# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      tags: [Documents]
      operationId: listDocuments
      summary: List published documents matching the curriculum mapping filters
      parameters:
        - name: yearLevel
          in: query
          required: false
          schema:
            type: integer
        - name: courseType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CourseType"
        - name: textId
          in: query
          required: false
          schema:
            type: integer
        - name: contentTypeId
          in: query
          required: false
          schema:
            type: integer
        - name: taskType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskType"
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Published documents matching the filters.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
  /documents/{slug}:
    get:
      tags: [Documents]
      operationId: getDocument
      summary: Get a published document by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/{slug}/progress:
    post:
      tags: [Documents]
      operationId: markDocumentViewed
      summary: Record that the current student viewed a document
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/{slug}/complete:
    post:
      tags: [Documents]
      operationId: markDocumentCompleted
      summary: Record that the current student completed a document
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Progress recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /admin/documents:
    get:
      tags: [Admin]
      operationId: adminListDocuments
      summary: List all documents including drafts (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/DocumentStatus"
        - name: contentTypeId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: All documents matching the filters.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Admin]
      operationId: adminCreateDocument
      summary: Create a new document (draft by default)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents/{id}:
    get:
      tags: [Admin]
      operationId: adminGetDocument
      summary: Get any document by id, including drafts (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [Admin]
      operationId: adminUpdateDocument
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Document updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents/{id}/publish:
    post:
      tags: [Admin]
      operationId: adminPublishDocument
      summary: Publish a document, snapshotting a new version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents/{id}/versions:
    get:
      tags: [Admin]
      operationId: adminListDocumentVersions
      summary: List version history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history, most recent first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/documents/{id}/versions/{versionId}/restore:
    post:
      tags: [Admin]
      operationId: adminRestoreDocumentVersion
      summary: Restore a document's body to a prior version (as a new draft)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document restored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "403":
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CourseType:
      type: string
      enum: [literature_atar, english_atar, general_english]
    TaskType:
      type: string
      enum: [essay, close_reading, short_answer, creative_writing]
    DocumentStatus:
      type: string
      enum: [draft, published]
...
    TextRef:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
    ContentType:
      type: object
      required: [id, slug, name, description]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
    DocumentSummary:
      type: object
      required:
        - id
        - title
        - slug
        - contentType
        - yearLevels
        - courseTypes
        - taskTypes
        - status
        - texts
        - publishedAt
        - updatedAt
        - isNew
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        contentType:
          $ref: "#/components/schemas/ContentType"
        yearLevels:
          type: array
          items:
            type: integer
        courseTypes:
          type: array
          items:
            $ref: "#/components/schemas/CourseType"
        taskTypes:
          type: array
          items:
            $ref: "#/components/schemas/TaskType"
        status:
          $ref: "#/components/schemas/DocumentStatus"
        texts:
          type: array
          items:
            $ref: "#/components/schemas/TextRef"
        publishedAt:
          type: [string, "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isNew:
          type: boolean
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/DocumentSummary"
        - type: object
          required: [body, copyrightNotice]
          properties:
            body:
              type: string
            copyrightNotice:
              type: string
    CreateDocumentBody:
      type: object
      required:
        - title
        - contentTypeId
        - body
        - yearLevels
        - courseTypes
        - taskTypes
        - textIds
      properties:
        title:
          type: string
          minLength: 1
        contentTypeId:
          type: integer
        body:
          type: string
        yearLevels:
          type: array
          items:
            type: integer
        courseTypes:
          type: array
          items:
            $ref: "#/components/schemas/CourseType"
        taskTypes:
          type: array
          items:
            $ref: "#/components/schemas/TaskType"
        textIds:
          type: array
          items:
            type: integer
        status:
          $ref: "#/components/schemas/DocumentStatus"
    UpdateDocumentBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        contentTypeId:
          type: integer
        body:
          type: string
        yearLevels:
          type: array
          items:
            type: integer
        courseTypes:
          type: array
          items:
            $ref: "#/components/schemas/CourseType"
        taskTypes:
          type: array
          items:
            $ref: "#/components/schemas/TaskType"
        textIds:
          type: array
          items:
            type: integer
        status:
          $ref: "#/components/schemas/DocumentStatus"
    DocumentVersion:
      type: object
      required: [id, documentId, versionNumber, body, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        versionNumber:
          type: integer
        body:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ProgressRecord:
      type: object
      required: [documentId, viewedAt, completedAt]
      properties:
        documentId:
          type: integer
        viewedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time