# lib/api-zod/src/generated/types/courseType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CourseType = typeof CourseType[keyof typeof CourseType];


export const CourseType = {
  literature_atar: 'literature_atar',
  english_atar: 'english_atar',
  general_english: 'general_english',
} as const;