# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentSummary } from './documentSummary';

export type DocumentDetail = DocumentSummary & {
  body: string;
  copyrightNotice: string;
};