# lib/api-zod/src/generated/types/documentVersion.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentVersion {
  id: number;
  documentId: number;
  versionNumber: number;
  body: string;
  createdAt: Date;
}