# lib/api-zod/src/generated/types/taskType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TaskType = typeof TaskType[keyof typeof TaskType];


export const TaskType = {
  essay: 'essay',
  close_reading: 'close_reading',
  short_answer: 'short_answer',
  creative_writing: 'creative_writing',
} as const;